making the search bar work with ui

+++ BookHeaven/Models/Book.cs
@@ -8,15 +8,15 @@
         public int bookId { get; set; }
         public string category { get; set; }
         public string format { get; set; }
-        public int price { get; set; }
+        public float price { get; set; }
         public int stock { get; set; }
         public string imageUrl { get; set; }
         public int ageLimitation { get; set; }
-        public int salePrice {  get; set; }
+        public float salePrice {  get; set; }
 
         public Book() { }
 
-        public Book(string name, string author, string date, int bookId, string category, string format, int price, int stock, string imageUrl, int ageLimitation, int salePrice)
+        public Book(string name, string author, string date, int bookId, string category, string format, float price, int stock, string imageUrl, int ageLimitation, float salePrice)
         {
             this.name = name;
             this.author = author;
